requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bitarray==3.1.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.1.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
cytoolz==1.0.1
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.3
django-oauth-toolkit==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
dotenv==0.9.9
drf-social-oauth2==3.1.0
drf-yasg==1.21.8
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
hexbytes==1.3.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
multidict==6.2.0
oauthlib==3.2.2
packaging==24.2
parsimonious==0.10.0
propcache==0.3.0
psycopg2-binary==2.9.10
# psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2024.2
pyunormalize==16.0.0
pywin32==310; sys_platform == "win32"
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
rpds-py==0.21.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
swagger-spec-validator==3.0.4
toolz==1.0.0
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2024.2
unique-namer==1.6.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
web3==7.9.0
websockets==13.1
whitenoise==6.9.0
yarl==1.18.3
